Refactor a couple non-specific to sentiment confusion matrix operations
